feat: Update redis
