+++ pkg/redis/src/index.ts
@@ -1,6 +1,6 @@
 import postgres from "postgres";
 import { config } from "dotenv";
-import { Redis } from "@upstash/redis";
+import { Redis } from "@upstash/redis/cloudflare";
 export function initRedis() {
   config();
   const redis = new Redis({
@@ -8,4 +8,10 @@ export function initRedis() {
     token: process.env.REDIS_TOKEN,
   });
   return redis;
+}
+export async function cacheJsonResults(redis: Redis, redisKey: string, returnObj: {}) {
+  const cachePipeline = redis.pipeline();
+  cachePipeline.json.set(redisKey, "$", returnObj);
+  cachePipeline.expire(redisKey, 600);
+  await cachePipeline.exec();
 }
